날짜 기준으로 알아서 몇회차인지 계산하도록 수정 -> 실행 할 때마다 오늘이 몇번째 꼬맨틀인지 바꿔주는거 귀찮음

+++ src/semantle/Semantle.java
@@ -8,7 +8,7 @@ import java.io.UnsupportedEncodingException;
 import java.util.List;
 
 public class Semantle {
-    private final int SEQ;
+    private final long SEQ;
     private final int DELAY;
     private final int PRINT_PER_TRY_COUNT;
 
@@ -16,7 +16,7 @@ public class Semantle {
     private static final String CORRECT = "정답!";
     private static final String NOT_FOUND = "찾지 못했습니다";
 
-    public Semantle(int SEQ, int DELAY, int PRINT_PER_TRY_COUNT) {
+    public Semantle(long SEQ, int DELAY, int PRINT_PER_TRY_COUNT) {
         this.SEQ = SEQ;
         this.DELAY = DELAY;
         this.PRINT_PER_TRY_COUNT = PRINT_PER_TRY_COUNT;
